// src/types/referral.ts
export interface ReferralFormData {
  referringProvider: string;
  referralContact: string;
  referralEmail: string;
  clientName: string;
  clientNumber: string;
  clientEmail: string;
  clientDOB: string;
  clientGender: string;
  clientState: string;
  insuranceInfo: string;
  referralPurpose: string;
}